Initial move to go std layout
