+++ internal/pkg/mapping/event_field.go
@@ -1,0 +1,8 @@
+package mapping
+
+type EventField int
+
+const (
+	Title EventField = iota
+	Description
+)
