+++ internal/pkg/mapping/range.go
@@ -1,0 +1,12 @@
+package mapping
+
+import "fmt"
+
+type Range struct {
+	StartRow int
+	EndRow   int
+}
+
+func (r *Range) String() string {
+	return fmt.Sprintf("%d:%d", r.StartRow, r.EndRow)
+}
